// types.ts
export enum ConversionType {
  REALISTIC_TO_CARTOON = 'realisticToCartoon',
  CARTOON_TO_REALISTIC = 'cartoonToRealistic',
}

export enum CartoonStyle {
  ANIME = 'anime',
  PIXAR = 'pixar',
  COMIC_BOOK = 'comicBook',
  CLASSIC_DISNEY = 'classicDisney'
}

export enum RealisticStyle {
  PHOTOREALISTIC = 'photorealistic',
  DIGITAL_PAINTING = 'digitalPainting',
  CINEMATIC = 'cinematic',
}